Added changes for compiling under VS.  Thanks 8DashP!
git-svn-id: http://scml-pp.googlecode.com/svn/trunk@73 f69afed9-71f0-c812-a9e8-761ac3947e42

+++ source/SCMLpp.cpp
@@ -8,6 +8,7 @@
 
 #ifndef _MSC_VER
     #include "libgen.h"
+    #include <algorithm>
 #endif
 
 #ifndef PATH_MAX
@@ -2507,8 +2508,8 @@ static void rotate_point(float& x, float& y, float angle, float origin_x, float 
     if(flipped)
         angle = -angle;
     
-    float s = sin(angle*M_PI/180);
-    float c = cos(angle*M_PI/180);
+    float s = sinf(angle*M_PI/180);
+    float c = cosf(angle*M_PI/180);
     float xnew = (x * c) - (y * s);
     float ynew = (x * s) + (y * c);
     xnew += origin_x;
